Merged my changes into Asset object
I managed to makde changes and forgot to rebase my repo. This is fixing that :) I've also added the full VERIS asset variety list, ownership and management type lists as at 01/27. And added the up to date relationships as described in TWIGs documentation.

One thing to note. I've removed the technical_characterization, as we have the evidenced-by relationship that will cover that off.