Initial fitout of the Identity Object with the fields of STIX v1.2
Having split the InformationSource object into Tool and Identity objects, this is the first cut of including the CIQ identity stuff, and the stix-resources information as discussed on the list. We WILL require further enhancement to include the CIQ identity fields. I suspect that we should probably just bite the bullet and include the important CIQ fields natively, as the OASIS CIL group have posted that they will not be updating the standard.